+++ index.html
@@ -41,6 +41,6 @@
             </div>
         </div>
         
-        <script src="src/index.js"></script>
+        <script type="module" src="src/index.js"></script>
     </body>
 </html>
